refactor(sqlite): split atomic.rs into atomic/ module directory
